+++ src/components/ReleaseNotes/ChangedSectionPreviewCard.tsx
